Faster world generation and fixed getting stuck if the max chunk distance was too small

+++ fastmath.cpp
@@ -19,7 +19,7 @@ void init_fastmath()
     float rad = 0;
     while(deg.value < LUT_SIZE)
 	{
-        table_sin[deg.value].fromFloat(sin(rad));
+        table_sin[deg.value].fromFloat(sinf(rad));
         deg.value++;
         rad += incr;
 	}
